fix(core): widen Signal.StartBit to ushort and fix decoder bit arithmetic
CAN FD payloads reach 64 bytes (512 bits), so a DBC Motorola signal
can have a StartBit in byte 32+ (value 256-511). The previous code
limited StartBit to byte (max 255) and used (byte)(start + i) inside
the decoders — both silently truncated, so any FD signal starting
past byte 31 was unparseable (parser threw OverflowException) or
mis-decoded (decoder read the wrong byte/offset pair).

Three coupled changes:

  1. Signal.StartBit: byte -> ushort. 64-byte CAN FD = 512 bits, so
     ushort is the natural choice (Signal.Length stays byte because
     the DBC spec caps signal width at 64 bits).

  2. SignalDecoder.ReadBigEndian and ReadLittleEndian: replace the
     (byte)(start + i) accumulator with int. byte truncation was the
     root cause of the mis-decode; the int accumulator naturally
     covers the full 0..511 range.

  3. DbcParser.ParseSignal: replace ParseByte(startTok) with a new
     ParseUShort helper. The old call wrapped any DBC with start
     bit >= 256 in a ParseFailure (the byte range error message).
     Only values past 65535 are now rejected, which is what the
     DBC spec actually requires.

Tests:
  - SignalDecoderTests.At_FD_Start_256/300/300 (3 new): Motorola
    and Intel decodes at start=256/300 in 64-byte payloads.
  - DbcParserTests.Parses_Signal_Start_Bit_Above_255 (1 new):
    start=300 in a 64-DLC message parses successfully.
  - DbcParserTests.Fails_On_Signal_Start_Bit_Overflow (updated):
    shifted the overflow boundary from 256 to 65536 to match the
    new ushort range. The old test asserted byte truncation
    behavior; the new one asserts ushort truncation.
  - DbcMessageViewModelTests: drop the (byte) cast on the loop
    index — it is no longer compatible with ushort StartBit.

+++ src/PeakCan.Host.Core/Dbc/Signal.cs
@@ -6,7 +6,9 @@ namespace PeakCan.Host.Core.Dbc;
 /// <see cref="MultiplexValue"/>) are populated by Task 6.
 /// </summary>
 /// <param name="Name">DBC signal identifier.</param>
-/// <param name="StartBit">0-based bit offset on the wire.</param>
+/// <param name="StartBit">0-based bit offset on the wire. Up to 511 for CAN FD
+/// (64-byte payload); ushort is required because byte (max 255) cannot
+/// represent a Motorola signal that starts past byte 31.</param>
 /// <param name="Length">Bit width. Up to 64 for CAN FD.</param>
 /// <param name="Order">Byte order on the wire.</param>
 /// <param name="ValueType">Numeric interpretation.</param>
@@ -22,7 +24,7 @@ namespace PeakCan.Host.Core.Dbc;
 /// <param name="ValueTableName">Optional <c>VAL_TABLE_</c> name attached via <c>VAL_</c> (Task 6).</param>
 public sealed record Signal(
     string Name,
-    byte StartBit,
+    ushort StartBit,
     byte Length,
     ByteOrder Order,
     ValueType ValueType,
